requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aioredis==1.3.1
    # via -r requirements.in
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
async-timeout==4.0.2
    # via aioredis
brotli==1.0.9
    # via yt-dlp
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
    #   yt-dlp
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.3.0
    # via email-validator
email-validator==2.0.0.post2
    # via fastapi
fastapi[all]==0.95.1
    # via -r requirements.in
ffmpeg-python==0.2.0
    # via -r requirements.in
future==0.18.3
    # via ffmpeg-python
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hiredis==2.2.2
    # via aioredis
httpcore==0.17.0
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.24.0
    # via fastapi
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iso8601==1.1.0
    # via m3u8
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
m3u8==3.4.0
    # via -r requirements.in
markupsafe==2.1.2
    # via jinja2
mutagen==1.46.0
    # via yt-dlp
orjson==3.8.10
    # via fastapi
pycryptodomex==3.17
    # via yt-dlp
pydantic==1.10.7
    # via fastapi
python-dotenv==1.0.0
    # via uvicorn
python-multipart==0.0.6
    # via fastapi
pyyaml==6.0
    # via
    #   fastapi
    #   uvicorn
requests==2.28.2
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via pydantic
ujson==5.7.0
    # via fastapi
urllib3==1.26.15
    # via requests
uvicorn[standard]==0.21.1
    # via fastapi
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.2
    # via
    #   uvicorn
    #   yt-dlp
yt-dlp==2023.3.4
    # via -r requirements.in
